% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{c_high}
\alias{c_high}
\title{correct cumulative, using high strategy}
\usage{
c_high(x)
}
\arguments{
\item{x}{vector of cumulative outcomes}
}
\description{
correct cumulative, using high strategy
}
\examples{
c_high(x)
}
\keyword{internal}
