% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{c_low}
\alias{c_low}
\title{correct cumulative, using low strategy}
\usage{
c_low(x)
}
\arguments{
\item{x}{vector of cumulative outcomes}
}
\description{
correct cumulative, using low strategy
}
\examples{
c_low(x)
}
\keyword{internal}
