% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{c_mid}
\alias{c_mid}
\title{correct cumulative, using mid strategy}
\usage{
c_mid(x)
}
\arguments{
\item{x}{vector of cumulative outcomes}
}
\description{
correct cumulative, using mid strategy
}
\examples{
c_mid(x)
}
\keyword{internal}
