% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{convert_weekly}
\alias{convert_weekly}
\title{Convert long csse data from daily to weekly format}
\usage{
convert_weekly(df, byvars = NULL)
}
\arguments{
\item{df}{this is the source long formatted data described above}

\item{byvars}{defaults to NULL, character vector of columns names to
group by}
}
\description{
This function will take a long format csse file (for example
something like that returned from \code{get_us_from_cdp(cdp)}, or
the long format returned from \code{cssedata()}, or the geo-level
elements in the list returney \code{us_empirical_by_level()}) and
convert it from daily data to weekly format.  The function will
use epiweek and epiyear from the lubridate package, and will
define the end of the week as day of week 7. The general approach
of the function is to sum up incident outcomes by epiweek/epiyear
combinations, and then generate weekly cumulative data by using
baseR \code{cumsum} function.
}
\examples{
convert_weekly(us)
convert_weekly(get_state_from_cdp("Texas",cdp), byvar="USPS")
convert_weekly(county, byvars="FIPS")
}
