% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{cssedata}
\alias{cssedata}
\title{pull data with improved speed}
\usage{
cssedata(gitpath = NULL, updategit = F)
}
\description{
This function pulls state level testing data from private and public labs, aggregated at state level
}
\examples{
dxtestingdata()
}
