% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{cssedataglobal}
\alias{cssedataglobal}
\title{Pull the Global CSSE data from the repo}
\usage{
cssedataglobal(gitpath = NULL, updategit = F)
}
\arguments{
\item{gitpath}{optional path to local repo}

\item{updategit}{option to update the local git repo}
}
\description{
This function pulls county level data from JHU CSSE repo (directly by defualt, from local repo if provided). Pulls Country Level Data
}
\examples{
cssedataglobal("jhudata/")
cssedataglobal()
}
