% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{dxtestingdata}
\alias{dxtestingdata}
\title{Pull the STATE level dx testing data from CDC}
\usage{
dxtestingdata()
}
\description{
This function pulls state level testing data from private and public labs, aggregated at state level
}
\examples{
dxtestingdata()
}
