% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{fast_pull}
\alias{fast_pull}
\title{pull data with improved speed}
\usage{
fast_pull()
}
\description{
This function pulls state level testing data from private and public labs, aggregated at state level
}
\examples{
dxtestingdata()
}
