% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{flexstate}
\alias{flexstate}
\title{Helper function to get state name wheter abbreviation or name is passed}
\usage{
flexstate(state)
}
\arguments{
\item{state}{state provided as either name or abbreviation}
}
\description{
Function returns statename given a statename in either abbrev or statename format
}
\examples{
flexstate("TX")
flexstate("Texas")
}
\keyword{internal}
