% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{get_urls}
\alias{get_urls}
\title{Function to update git and finalize urls}
\usage{
get_urls(gitpath = NULL, updategit = F, scope = c("US", "global"))
}
\arguments{
\item{gitpath}{string path to local git repo, default NULL}

\item{updategit}{boolean default FALSE, set to TRUE to update the repo}

\item{scope}{= string, one of "US" or "global"}
}
\description{
Function takes gitpath or NULL, a flag to update the git
and a scope
}
\examples{
get_urls(scope="US")
get_urls(scope="global")
get_urls(gitpath="mypath/", updategit=T, scope="US")
}
