% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{get_us_from_cdp}
\alias{get_us_from_cdp}
\title{Function to just get a US from c,d,p}
\usage{
get_us_from_cdp(c, d, p)
}
\arguments{
\item{c}{compact version of confirmed cases}

\item{d}{compact version of deaths}

\item{p}{compact version of population}
}
\description{
Function use compact csse data (c,d,p) to return
a datatable for entire US
}
\examples{
get_us_from_cdp(c,d,p)
}
