% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_covid}
\alias{plot_covid}
\title{Function to plot a location}
\usage{
plot_covid(dt, ...)
}
\arguments{
\item{dt}{data frame containing the data}

\item{...}{parameters (as above) passed on}
}
\value{
ggplot2 plot
}
\description{
Function returns plot of a given location. The only required input is
a data frame containing covid data to be plotted. Without any other
arguments, the function will look for an xvar named "Date", an outcome
variable named "Confirmed", and will smooth these using a generalized additive
model with natural cubic spline regression (following \code{mgcv::gam}, with \code{bs="cs"})
and knots placed every 21 days, and poisson distribution assumed for the outcome
These parameters can be adjusted by the user, including passing along user-defined
values for \code{xvar} (for example "date"), \code{yvar} (e.g. "Deaths"), \code{knot_interval} (e.g.
7), and \code{family} (e.g. "gaussian")
}
\examples{
plot_covid(brazil)
plot_covid(australia, yvar="Deaths", knot_interval=7, family="gaussian")
}
