% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_single_location}
\alias{plot_single_location}
\title{Plot single location}
\usage{
plot_single_location(dt, sm = "phat", xvar = "Date", yvar = "Confirmed", ...)
}
\arguments{
\item{dt}{data frame containing the data to be plotted}

\item{sm}{string name of column in \code{dt} containing smoothed observations}

\item{xvar}{string name of column in \code{dt} containing xvar, default is "Date"}

\item{yvar}{string name of column in \code{dt} containng yvar, default is "Confirmed"}
}
\value{
ggplot2 object
}
\description{
Function receives data frame/tibble/dt, and names of smoothed
the empirical data and the smoothed outcome, xvar, and yvar
and returns a plot
}
\examples{
plot_single_location(brazil)
plot_single_location(us,yvar="Deaths")
}
\keyword{internal}
