% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{smooth_dt}
\alias{smooth_dt}
\title{Return smooth predictions, using generalized additive model}
\usage{
smooth_dt(x, y, k, family = c("poisson", "gaussian"))
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}

\item{k}{integer number of knots}

\item{family}{string indicating family, must be either "poisson" (default)
or "gaussian"}
}
\value{
vector of smoothed predictions
}
\description{
Function returns the predictions from a simple generalized
additive model that uses natural cubic splines to regress
y on splines of x, with k knots, given the family
}
\examples{
smooth_dt(x,y,10,"poisson")
}
\keyword{internal}
