% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{smooth_pre_process}
\alias{smooth_pre_process}
\title{Pre process smoothing}
\usage{
smooth_pre_process(
  dt,
  xvar = "Date",
  yvar = "Confirmed",
  knot_interval = 21,
  family = c("poisson", "gaussian")
)
}
\arguments{
\item{dt}{data frame containing the data to be smoothed}

\item{xvar}{string name of column in \code{dt} containing xvar, default is "Date"}

\item{yvar}{string name of column in \code{dt} containng yvar, default is "Confirmed"}

\item{knot_interval}{how frequently should the nots be placed? default is 21,
which in the case of plotting daily covid data, represents one knot every 3 weeks}

\item{family}{string indicating family, must be either "poisson" (default)
or "gaussian"}
}
\value{
named list containing x,y,and k, which can be fed to \code{smooth_dt}
}
\description{
Function pre-processes data for sending to smoothing function
}
\examples{
smooth_pre_process(brazil)
}
\keyword{internal}
