% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{statenames}
\alias{statenames}
\title{Prepare a utility dataframe for state names and abbreviations}
\usage{
statenames()
}
\description{
This function prepares a utility dataframe for state names and abbreviations
}
\examples{
statenames()
}
