% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{us_empirical_by_level}
\alias{us_empirical_by_level}
\title{Prepare us level empirical data into three levels}
\usage{
us_empirical_by_level(
  empirical_source,
  csse_repo_path = NULL,
  filterdates = NULL,
  updategit = F,
  supplement_territories = T
)
}
\arguments{
\item{empirical_source}{source of empirical data.. must be 'csse','usafacts'}

\item{csse_repo_path}{defaults to NULL, otherwise a path to local clone of JHU CSSE repo}

\item{filterdates}{defaults to all NULL, otherwise a pair of dates (e.g. c("2020-03-01", "2020-06-01")}

\item{updategit}{defaults to FALSE; should the local csse git repo be queried for possible update?}
}
\description{
This function prepares the empirical data. Depending on source it will call functions to prepare the csse or usafacts data
}
\examples{
prepare_empirical_data("usafacts", "jhudata/", c("2020-03-01", "2020-06-01"))
prepare_empirical_data("csse")
}
