% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{cssedataglobal}
\alias{cssedataglobal}
\title{Function to pull global data quickly}
\usage{
cssedataglobal(gitpath = NULL, updategit = F)
}
\description{
Function is faster approach to returning global data
}
