% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{get_all_states_from_cdp}
\alias{get_all_states_from_cdp}
\title{Get all states from cdp}
\usage{
get_all_states_from_cdp(cdp)
}
\arguments{
\item{cdp}{this is a cdp list as returned by \code{cssedata(return_compact=T)}}
}
\description{
This function extends the approach of \code{get_state_from_cdp()}, returining
all states in long format, rather than a single state
}
\examples{
get_all_states_from_cdp(cdp)
}
