% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{get_county_from_cdp}
\alias{get_county_from_cdp}
\title{Function to just get a US county from c,d,p}
\usage{
get_county_from_cdp(fips, cdp)
}
\arguments{
\item{cdp}{compact version of cases, deaths, population as return from
cssedata(return_compact=T)}
}
\description{
Function use compact csse data (c,d,p) to return
a datatable for entire US
}
\examples{
get_county_from_cdp(21027,cdp)
}
