% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{get_state_from_cdp}
\alias{get_state_from_cdp}
\title{Function to just get a single state given cdp}
\usage{
get_state_from_cdp(state, cdp)
}
\arguments{
\item{state}{string}

\item{cdp}{compact version of cases, deaths, population as return from
cssedata(return_compact=T)}
}
\description{
Function use compact csse data (c,d,p) to return
a datatable for a specific state only
}
\examples{
get_state_from_cdp("Maryland", cdp)
}
