% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{get_state_from_cdp}
\alias{get_state_from_cdp}
\title{Function to just get a single state given cdp}
\usage{
get_state_from_cdp(
  cdp,
  state = NULL,
  fix_cumul = FALSE,
  type = c("mid", "low", "high")
)
}
\arguments{
\item{cdp}{compact version of cases, deaths, population as return from
cssedata(return_compact=T)}

\item{state}{string (default is NULL, will return all states), otherwise specifiy a
specifc state abbreviation (i.e. "TX") or state name (i.e. "Texas")}

\item{fix_cumul}{(logical, default=F); set to TRUE to fix cumulative data}

\item{type}{(string, default="mid"); strategy to fix cumulative data}
}
\description{
Function use compact csse data (c,d,p) to return
a datatable for a specific state only
}
\examples{
get_state_from_cdp(cdp, "MD")
}
