% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pull_functions.R
\name{quick_melt}
\alias{quick_melt}
\title{Function to quickly melt c an d into a data table}
\usage{
quick_melt(c, d)
}
\arguments{
\item{c}{compact version of confirmed cases}

\item{d}{compact version of deaths}
}
\description{
Given a compact version of cases and deaths, function
will convert the information into a datatable
first summing over all the rows of c and d
}
\examples{
quick_melt(c,d)
}
